Refactor around switch statement
