+++ includes/class-static-press-ajax-finalyze.php
@@ -10,11 +10,19 @@ namespace static_press\includes;
 if ( ! class_exists( 'static_press\includes\Static_Press_Ajax_Finalyze' ) ) {
 	require dirname( __FILE__ ) . '/class-static-press-ajax_finalyze.php';
 }
+if ( ! class_exists( 'static_press\includes\Static_Press_Response_Processor_200' ) ) {
+	require dirname( __FILE__ ) . '/class-static-press-response-processor-200.php';
+}
+if ( ! class_exists( 'static_press\includes\Static_Press_Response_Processor_404_Dump' ) ) {
+	require dirname( __FILE__ ) . '/class-static-press-response-processor-404-dump.php';
+}
 if ( ! class_exists( 'static_press\includes\Static_Press_Transient_Manager' ) ) {
 	require dirname( __FILE__ ) . '/class-static-press-transient-manager.php';
 }
 
 use static_press\includes\Static_Press_Ajax_Processor;
+use static_press\includes\Static_Press_Response_Processor_200;
+use static_press\includes\Static_Press_Response_Processor_404_Dump;
 use static_press\includes\Static_Press_Transient_Manager;
 
 /**
@@ -25,7 +33,11 @@ class Static_Press_Ajax_Finalyze extends Static_Press_Ajax_Processor {
 	 * Creates 404 error page html static file, 
 	 */
 	protected function process_ajax_request() {
-		$this->create_static_file( Static_Press_Url_Collector::get_site_url() . '404.html' );
+		$static_file_creator = $this->create_static_file_creator_remote(
+			new Static_Press_Response_Processor_200(),
+			new Static_Press_Response_Processor_404_Dump()
+		);
+		$static_file_creator->create( Static_Press_Url_Collector::get_site_url() . '404.html' );
 		Static_Press_Transient_Manager::delete_transient();
 
 		$result = array( 'result' => true );
